refactor: isolate guided profile conversation lifecycle
Changed:
Moved submission parsing and stream event interpretation into a pure model, and centralized session creation, cursor replay, retries, aborts, and terminal handling in a dedicated controller hook.

Benefit:
The MarkdownFlow component now focuses on rendering while preserving the existing guided conversation behavior and public interface.
